+++ src/App.vue
@@ -78,7 +78,7 @@ export default {
     // browsers do not display this message, but provide a generic
     // "content has changed" dialog instead.
     $(window).on('beforeunload', () => {
-      const confirmationMessage = 'Your modifications have not been saved and will be lost if you close the Curation Tool. Confirm to discard your changes, or cancel to return to the Curation Tool.';
+      const confirmationMessage = 'Your modifications have not been saved and will be lost if you close Klados. Confirm to discard your changes, or cancel to return to Klados.';
 
       if (!isEqual(this.loadedPhyx, this.currentPhyx)) return confirmationMessage;
       return false;