+++ vue.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   // Set a prefix so we can host this on Github Pages.
-  publicPath: '/curation-tool/',
+  publicPath: '/klados/',
 
   // Write out to ./docs so we can host on Github Pages without
   // deploying separately to gh-pages.