+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# Change log for the Curation tool
+# Change log for Klados
 
 Based on the suggestion at https://keepachangelog.com/en/1.0.0/.
 
 ## [Unreleased]
+- Rename Curation Tool to Klados ([#161](https://github.com/phyloref/curation-tool/issues/161))
 - Rewrote Curation Tool using Vue CLI to break it up into components.
 
 ## 0.1.0 - 2018-05-21